+++ src/core/options.ts
@@ -1,0 +1,1 @@
+export type Options = any;
